+++ libs/core/src/lib/adapters/create-noop-reaction.function.ts
@@ -1,0 +1,8 @@
+export function createNoopReaction<State>(): NoopReaction<State> {
+  return (state: State) => state;
+}
+
+export type NoopReaction<State> = (s: State) => State;
+export interface WithNoopReaction<State> {
+  noop: NoopReaction<State>;
+}